requirements.txt
Django==1.9.4
djangorestframework==3.3.3
Pillow==3.1.1
